Fixed terrifying string misuses, fixed statusLED bug
There were several places in code where char* were being assigned values, which is NOT how that works in C++, and can cause "undefined behavior", which it did. I re-wrote all places where char* was being used incorrectly, which fixed an issue where sometimes the MACs of other Lumas in the group (and potentially other parts of memory) were being set to null or giberish, causing all sorts of bad stuff to happen.

I also added a check so that if a message is received from a client that is beleived to be offline, it re-checks who's online to ensure correct status LED indication.